Set StatusBar height from insets
