+++ lib/android/app/src/main/java/com/reactnativenavigation/utils/UiUtils.java
@@ -2,7 +2,6 @@ package com.reactnativenavigation.utils;
 
 import android.content.Context;
 import android.content.res.Resources;
-import android.os.Build;
 import android.os.Handler;
 import android.os.Looper;
 import android.support.annotation.NonNull;
@@ -13,11 +12,8 @@ import android.view.ViewTreeObserver;
 import android.view.WindowManager;
 
 public class UiUtils {
-    private static final int STATUS_BAR_HEIGHT_M = 24;
-    private static final int STATUS_BAR_HEIGHT_L = 25;
     private static final int DEFAULT_TOOLBAR_HEIGHT = 56;
 
-    private static int statusBarHeight = -1;
     private static int topBarHeight = -1;
 
     public static <T extends View> void runOnPreDrawOnce(@Nullable final T view, final Functions.Func1<T> task) {
@@ -73,18 +69,6 @@ public class UiUtils {
             windowManager.getDefaultDisplay().getMetrics(metrics);
         }
         return metrics;
-    }
-
-    public static int getStatusBarHeight(Context context) {
-        if (statusBarHeight > 0) {
-            return statusBarHeight;
-        }
-        final Resources resources = context.getResources();
-        final int resourceId = resources.getIdentifier("status_bar_height", "dimen", "android");
-        statusBarHeight = resourceId > 0 ?
-                resources.getDimensionPixelSize(resourceId) :
-                dpToPx(context, Build.VERSION.SDK_INT >= Build.VERSION_CODES.M ? STATUS_BAR_HEIGHT_M : STATUS_BAR_HEIGHT_L);
-        return statusBarHeight;
     }
 
     public static int getTopBarHeightDp(Context context) {
