+++ lib/android/app/src/main/java/com/reactnativenavigation/viewcontrollers/ChildController.java
@@ -8,6 +8,7 @@ import android.view.ViewGroup;
 
 import com.reactnativenavigation.parse.Options;
 import com.reactnativenavigation.presentation.Presenter;
+import com.reactnativenavigation.utils.StatusBarUtils;
 import com.reactnativenavigation.viewcontrollers.navigator.Navigator;
 import com.reactnativenavigation.views.Component;
 
@@ -87,6 +88,7 @@ public abstract class ChildController<T extends ViewGroup> extends ViewControlle
     }
 
     private WindowInsetsCompat applyWindowInsets(WindowInsetsCompat insets) {
+        StatusBarUtils.setStatusBarHeight(insets.getSystemWindowInsetTop());
         return insets;
     }
 }
